% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectRversion.R
\name{computeLatentSpace}
\alias{computeLatentSpace}
\title{function to output the latentspace determined by the autoencoder.}
\usage{
computeLatentSpace(ods)
}
\arguments{
\item{ods}{An OUTRIDER data set}
}
\description{
function to output the latentspace determined by the autoencoder.
}
