% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValMatrix.R
\docType{methods}
\name{computePvalues}
\alias{computePvalues}
\alias{computePvalues,OutriderDataSet-method}
\alias{computePvalues,OutriderDataSet-method}
\title{This is the main function to calculate all P-values.}
\usage{
computePvalues(object, ...)

\S4method{computePvalues}{OutriderDataSet}(object,
  alternative = c("two.sided", "greater", "less"), method = "BH",
  modelFile = NULL, BPPARAM = bpparam())
}
\arguments{
\item{object}{object}

\item{alternative}{"two.sided", "greater" or "less" P-values can be computed, default is 
"two.sided"}

\item{method}{method used for multiple testing}

\item{BPPARAM}{by default bpparam()}

\item{by}{direction of multiple testing correction, by default: sample.}
}
\value{
a OutriderDataSet object with computed pValues and padjust as an assay
}
\description{
Computes two matrices with the same dimension as the count matrix, 
which contain the corresponding P-values and adjusted P-values to every 
count.
}
\examples{
ods <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods <- estimateSizeFactors(ods)
ods <- fit(ods)
ods <- computePvalues(ods)

assays(ods)[['pValue']][1:10,1:10]

ods <- computeZscores(ods)
res <- results(ods)
res

}
\seealso{
p.adjust
}
