% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZscoreMatrix.R
\docType{methods}
\name{computeZscores}
\alias{computeZscores}
\alias{computeZscores,OutriderDataSet-method}
\title{Z score matrix}
\usage{
computeZscores(ods, ...)

\S4method{computeZscores}{OutriderDataSet}(ods, ...)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{...}{further arguments passed to \code{ZscoreMatrix}}
}
\value{
matrix of z scores and l2fc
}
\description{
computes the z scores for every log2 fold-change compared
to the mean(normalizedCounts)
}
\examples{
ods <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods <- estimateSizeFactors(ods)
ods <- computeZscores(ods)

assays(ods)[['zScore']][1:10,1:10]
assays(ods)[["l2fc"]][1:10,1:10]

}
