% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,OutriderDataSet-method}
\alias{counts<-,OutriderDataSet,matrix-method}
\alias{counts<-,OutriderDataSet,matrix-method}
\title{Accessors for the 'counts' slot of an OutriderDataSet object.}
\usage{
\S4method{counts}{OutriderDataSet}(object, normalized = FALSE, offset = 0)

\S4method{counts}{OutriderDataSet,matrix}(object) <- value
}
\arguments{
\item{object}{OutriderDataSet}

\item{normalized}{TRUE/FALSE whether counts should be normalized}

\item{offset}{Pseudocount offset by default 0.}

\item{value}{An integer matrix containing the counts}
}
\value{
A matrix containing the counts
}
\description{
The counts slot holds the count data as a matrix of non-negative integer
count values, one row for each observational unit (eg.: gene), and one
column for each sample.
}
\details{
By default this function returns the raw counts. If conrol factors are
computed or provided the normalized counts can be returned using 
normalized = TRUE. The offset parameter can be used to add a pseudocount
to the count before dividing by the normalization. This can be usefull 
when the log(counts) are computed and in case the controll values are in 
the same oder of magnited as the counts.
}
\examples{

ods <- makeExampleOutriderDataSet()
counts(ods)[1:10,1:10]

ods <- estimateSizeFactors(ods)
counts(ods, normalized=TRUE)[1:10,1:10]

}
\seealso{
\code{\link{sizeFactors}}, \code{\link{normalizationFactors}}
}
