% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNSetterFuns.R
\name{default_param_getters}
\alias{default_param_getters}
\alias{getDefaultPreproParams,}
\alias{getDefaultPvalueParams}
\alias{getDefaultPreproParams}
\title{Default paramenter getter functions}
\usage{
getDefaultPreproParams(ods)

getDefaultPvalueParams(ods)
}
\arguments{
\item{ods}{An OutriderDataSet object.}

\item{...}{Further arguments passed on to the underlying assay function.}
}
\value{
A list of parameters.
}
\description{
This is a collection of functions for easy access to the default parameter 
settings within the OUTRIDER2 model based on the profile specifed in the 
Outrider2DataSet. If needed, the user can adapt the parameters and pass 
them on to the OUTRIDER function.
}
\examples{
ods <- makeExampleOutriderDataSet(10, 10)
getDefaultPreproParams()
getDefaultPvalueParams()


}
