% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{estimateBestQ}
\alias{estimateBestQ}
\title{Estimation of Q}
\usage{
estimateBestQ(ods)
}
\arguments{
\item{ods}{An OutriderDataSet object}
}
\value{
The estimated dimension of hidden confounders
}
\description{
Estimating the best q for the given data set
}
\examples{
ods <- makeExampleOutriderDataSet()

estimateBestQ(ods)

}
