% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNB.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,OutriderDataSet-method}
\alias{fit,OutriderDataSet-method}
\title{Fit a NB distribution to the genecounts per gene over all samples using the
precomputed control factors.
Parallized version.}
\usage{
fit(object, ...)

\S4method{fit}{OutriderDataSet}(object, dropExtremeRank = FALSE,
  modelFile = NULL, BPPARAM = bpparam())
}
\arguments{
\item{object}{OutriderDataSet}

\item{...}{additional arguments.}

\item{modelFile}{file name to save the estimated parameters}

\item{BPPARAM}{by default bpparam()}
}
\description{
Fit a NB distribution to the genecounts per gene over all samples using the
precomputed control factors.
Parallized version.
}
\examples{
ods <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods <- estimateSizeFactors(ods)
ods <- fit(ods)
  
mcols(ods)[1:10,c('mu', 'disp')]
  
}
