% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectEDversion.R
\name{fitAutoencoder}
\alias{fitAutoencoder}
\title{Main autoencoder fit function}
\usage{
fitAutoencoder(ods, q, thetaRange = c(0.01, 1000), convergence = 1e-05,
  iterations = 15, initialize = TRUE, correctTheta = "none",
  usePCA = TRUE, lasso = FALSE, runLassoFit = TRUE, nFolds = 10,
  control = list(), useOptim = TRUE, L1encoder = FALSE,
  BPPARAM = bpparam())
}
\description{
initialize D
}
