% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getCovariatesOneHotEncoded}
\alias{getCovariatesOneHotEncoded}
\title{Get one hot encoding for specified covariates}
\usage{
getCovariatesOneHotEncoded(ods)
}
\description{
Get one hot encoding for specified covariates
}
