% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-gridSearch.R
\name{injectOutliers}
\alias{injectOutliers}
\title{injectOutliers}
\usage{
injectOutliers(ods, freq, zScore, inj)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{freq}{frequency of injected counts.}

\item{zScore}{injection Z-score.}

\item{inj}{injection strategy.}
}
\value{
and OutriderDataSet with artificially corrupted counts.
}
\description{
injectOutliers
}
