% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectRversion.R
\name{loss}
\alias{loss}
\title{loss function}
\usage{
loss(w, k, x, s, xbar, theta)
}
\arguments{
\item{w}{weight matrix}

\item{k}{counts}

\item{x}{log-centered counts}

\item{s}{size factors}

\item{xbar}{offset}

\item{theta}{value used in the likelihood (default=25).}
}
\value{
Returns the negative log likelihood of the negative binomial
}
\description{
loss function
}
