% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectRversion.R
\name{lossGrad}
\alias{lossGrad}
\title{gradient of loss function}
\usage{
lossGrad(w, k, x, s, xbar, theta)
}
\arguments{
\item{w}{weight matrix}

\item{k}{counts}

\item{x}{log-centered counts}

\item{s}{size factors}

\item{xbar}{offset}

\item{theta}{value used in the likelihood (default=25).}
}
\value{
returns the gradient of the loss function.
}
\description{
gradient of loss function
}
