% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Outrider2DataSet.R
\name{makeExampleProtriderDataSet}
\alias{makeExampleProtriderDataSet}
\title{Create example data sets for OUTRIDER with protein intensities}
\usage{
makeExampleProtriderDataSet(
  n = 200,
  m = 80,
  q = 10,
  freq = 0.001,
  zScore = 6,
  inj = c("both", "low", "high"),
  sf = rnorm(m, mean = 1, sd = 0.1)
)
}
\arguments{
\item{n}{Number of simulated proteins}

\item{m}{Number of simulated samples}

\item{q}{number of simulated latend variables.}

\item{freq}{Frequency of in-silico outliers}

\item{zScore}{Absolute z score of in-silico outliers (default 6).}

\item{inj}{Determines whether outliers are injected with the strategy 
('both', 'low', 'high'), default is 'both'.}

\item{sf}{Artificial Size Factors}
}
\value{
An ProtriderDataSet containing an example dataset
}
\description{
Creates an example data set by simulating a data set based 
on random intensities following a log-normal distribution with injected
outliers with a fixed z score away from the mean of the gene.
}
\examples{
# A generic dataset 
pds <- makeExampleProtriderDataSet()
pds

}
