% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{normalizationFactors}
\alias{normalizationFactors}
\alias{normalizationFactors,OutriderDataSet-method}
\alias{normalizationFactors<-,OutriderDataSet,matrix-method}
\alias{normalizationFactors<-,OutriderDataSet,DataFrame-method}
\alias{normalizationFactors<-,OutriderDataSet,NULL-method}
\alias{normalizationFactors<-,OutriderDataSet,matrix-method}
\alias{normalizationFactors<-,OutriderDataSet,DataFrame-method}
\alias{normalizationFactors<-,OutriderDataSet,data.frame-method}
\alias{normalizationFactors<-,OutriderDataSet,NULL-method}
\title{Accessor functions for the normalization factors in a OutriderDataSet
object.}
\usage{
\S4method{normalizationFactors}{OutriderDataSet}(object)

\S4method{normalizationFactors}{OutriderDataSet,matrix}(object,
  replace = TRUE) <- value

\S4method{normalizationFactors}{OutriderDataSet,DataFrame}(object,
  replace = TRUE) <- value

\S4method{normalizationFactors}{OutriderDataSet,data.frame}(object,
  replace = TRUE) <- value

\S4method{normalizationFactors}{OutriderDataSet,`NULL`}(object) <- value
}
\arguments{
\item{object}{a \code{OutriderDataSet} object.}

\item{value}{the matrix of normalization factors}

\item{replace}{if old values are present values are replcaed. If set to false
old and new values are multiplied.}

\item{...}{additional arguments}
}
\description{
Accessor functions for the normalization factors in a OutriderDataSet
object.
}
\examples{

ods <- makeExampleOutriderDataSet(n=111, m=10)

normFactors <- matrix(runif(nrow(ods)*ncol(ods),0.5,1.5),
    ncol=ncol(ods),nrow=nrow(ods))

# the normalization factors matrix should not have 0's in it
# it should have geometric mean near 1 for each row
normFactorsRM <- normFactors / exp(rowMeans(log(normFactors)))
normalizationFactors(ods) <- normFactorsRM
normalizationFactors(ods)[1:10,1:10]

normalizationFactors(ods) <- NULL
ods <- estimateSizeFactors(ods)
normalizationFactors(ods, replace=FALSE) <- normFactors
all(normalizationFactors(ods) == t(sizeFactors(ods) * t(normFactors)))

}
\seealso{
DESeq2::normalizationFactors
}
