% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Outrider2DataSet.R
\docType{class}
\name{Outrider2DataSet-class}
\alias{Outrider2DataSet-class}
\alias{Outrider2DataSet}
\title{Outrider2DataSet class and constructors}
\usage{
Outrider2DataSet(
  se,
  inputData,
  colData,
  distribution = c("gaussian", "negative binomial"),
  sf_norm = TRUE,
  preprocessing = c("log", "vst", "none"),
  transformation = c("none", "log"),
  ...
)
}
\arguments{
\item{se}{A RangedSummarizedExperiment object or any object which inherits
from it and contains a count matrix as the first element in the
assay list.}

\item{colData}{Additional to the count data a \code{DataFrame} can be 
provided to annotate the samples.}

\item{...}{Further arguments can be passed to
\code{\link[DESeq2]{DESeqDataSet}}, which is used to parse the
user input and create the initial 
\code{RangedSummarizedExperiment} object.}

\item{countData}{A simple count matrix. If dim names are provided, they have
to be unique. This is only used if no \code{se} object is
provided.}
}
\value{
An Outrider2DataSet object.
}
\description{
The Outrider2DataSet class is designed to store the whole 
OUTRIDER data set needed for an analysis. It is a subclass of 
\code{RangedSummarizedExperiment}. All calculated values and results are 
stored as assays or as annotation in the mcols structure provided by the
\code{RangedSummarizedExperiment} class.
}
\examples{

ods <- makeExampleOutrider2DataSet()
ods

ods <- makeExampleOutriderDataSet()
ods
   
}
\author{
Christian Mertes \email{mertes@in.tum.de}, 
            Felix Brechtmann \email{brechtma@in.tum.de}
}
