% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OutriderDataSet.R
\docType{class}
\name{OutriderDataSet-class}
\alias{OutriderDataSet-class}
\title{OutriderDataSet}
\description{
This class is designed to store the whole OUTRIDER data set
needed for an analysis of a disease cohort
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
