% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OutriderDataSet.R
\name{OutriderDataSet}
\alias{OutriderDataSet}
\title{The constructor function for OutriderSettings}
\usage{
OutriderDataSet(se, countData, colData, ...)
}
\arguments{
\item{se}{RangedSummarizedExperiment object}

\item{countData}{countData}

\item{colData}{additional Annotation Data}

\item{...}{Any parameters corresponding to the slots and their possible
values. See \linkS4class{OutriderDataSet}}
}
\value{
A OutriderDataSet object.
}
\description{
Eather a RangedSummarizedExperiment object or a count table has to be 
provided.
}
\examples{
    ods <- makeExampleOutriderDataSet()
    ods
    
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
