% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValMatrix.R
\name{padjMatrix}
\alias{padjMatrix}
\title{FDR correction}
\usage{
padjMatrix(ods, method = "BH")
}
\arguments{
\item{ods}{OutriderDataSet}

\item{method}{adjustment method, by default 'BH'}
}
\value{
OutriderDataSet containing adjusted pValues
}
\description{
Function to correct the computed pValues using BY FDR correction.
This function is called by the computePvalues method.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)
ods <- padjMatrix(ods, method='BH')
head(assays(ods)[["padjust"]][,1:10])

ods <- padjMatrix(ods, method='fdr')
head(assays(ods[,1:10])[["padjust"]])

}
