% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotCountCorHeatmap}
\alias{plotCountCorHeatmap}
\title{Correlation heatmap of the count data of the given samples}
\usage{
plotCountCorHeatmap(x, normalized = TRUE, rowCoFactor = NULL,
  rowColSet = "Set1", colCoFactor = NULL, colColSet = "Set2",
  nCluster = 4, main = "Count correlation heatmap",
  annotateCluster = TRUE, dendrogram = "both", names = c("both", "row",
  "col", "none"), ...)
}
\value{
OutriderDataSet object
}
\description{
Correlation heatmap of the count data of the given samples
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- plotCountCorHeatmap(ods, annotateCluster=TRUE)
  
sex <- sample(c("female", "male"), dim(ods)[2], replace=TRUE)
colData(ods)$sex <- sex 
plotCountCorHeatmap(ods, colCoFactor="sex")
  
}
