% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\docType{methods}
\name{plotDispEsts,OutriderDataSet-method}
\alias{plotDispEsts,OutriderDataSet-method}
\title{Dispersion estimation plot}
\usage{
\S4method{plotDispEsts}{OutriderDataSet}(object)
}
\description{
Plotting the dispersion of the OutriderDataSet model against the normalized 
mean count.
}
