% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotExpressionRank}
\alias{plotExpressionRank}
\title{Plot expression over expression rank per gene.
The plot can be used before and after fitting.}
\usage{
plotExpressionRank(ods, geneID, padj = 0.05, basePlot = FALSE)
}
\arguments{
\item{ods}{A OUTRIDER data set.}

\item{geneID}{Id of gene wich should be plotted. 
If multiple gene ids are supplied mulitple plots will be made.}

\item{padj}{padj cutoff for coloring significant genes}

\item{basePlot}{R base plot version.}
}
\description{
Plot expression over expression rank per gene.
The plot can be used before and after fitting.
}
\examples{
ods <- makeExampleOutriderDataSet(10, 100)
ods <- OUTRIDER(ods)
plotExpressionRank(ods, 1)

}
