% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotFPKM}
\alias{plotFPKM}
\title{Distribution plot of FPKM values}
\usage{
plotFPKM(ods)
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}
}
\value{
A ggplot object containing the FPKM plot
}
\description{
Plots the distribution of FPKM values. It also can compare between
filtered and non filtered features.
}
\examples{
ctsFile <- system.file('extdata', 'GTExSkinSmall.tsv', package='OUTRIDER')
ods <- OutriderDataSet(countData=read.table(ctsFile, check.names=FALSE))
annotation <- system.file('extdata', 'gencode.v19.genes.small.gtf.gz', 
        package='OUTRIDER')

# Filter, storing the fpkm values and not subsetting the ods object.
ods <- filterExpression(ods, annotation, filterGenes=FALSE, savefpkm=TRUE)

#Display the pre distribution of counts.
plotFPKM(ods)

}
