% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotFunctions}
\alias{plotQQ}
\alias{plotExpressionRank}
\alias{plotCountCorHeatmap}
\alias{plotAberrantPerSample}
\alias{plotFPKM}
\alias{plotDispEsts}
\alias{plotPowerAnalysis}
\alias{plotQQ}
\alias{plotExpressionRank}
\alias{plotCountCorHeatmap}
\title{Visualization functions for OUTRIDER}
\usage{
plotVolcano(ods, sampleID, main, padjCutoff = 0.05, zScoreCutoff = 0,
  pch = 16, basePlot = FALSE, col = c("gray", "firebrick"))

plotQQ(ods, geneID, main, global = FALSE, padjCutoff = 0.05,
  zScoreCutoff = 0, samplePoints = TRUE, legendPos = "topleft",
  outlierRatio = 0.001, conf.alpha = 0.05, pch = 16, xlim = NULL,
  ylim = NULL, col = NULL)

plotExpressionRank(ods, geneID, main, padjCutoff = 0.05, zScoreCutoff = 0,
  normalized = TRUE, basePlot = FALSE, log = TRUE, col = c("gray",
  "firebrick"), ...)

plotCountCorHeatmap(ods, normalized = TRUE, rowCentered = TRUE,
  rowCoFactor = NULL, rowColSet = "Set1", colCoFactor = NULL,
  colColSet = "Set2", nCluster = 4, main = "Count correlation heatmap",
  annotateCluster = TRUE, dendrogram = "both", basePlot = TRUE,
  names = c("both", "row", "col", "none"), ...)
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}

\item{sampleID}{A sampleID, which should be plotted. Can also be a vector.}

\item{main}{Title for the plot, if missing a default title will be used.}

\item{padjCutoff}{Significance level to mark outliers}

\item{zScoreCutoff}{Z-score cutoff to mark outliers}

\item{pch}{Integer or character to be used for plotting the points}

\item{basePlot}{if TRUE, use the R base plot version, else use the plotly 
framework, which is the default}

\item{col}{Set color for the points. If set, it must be a character vector 
of length 2. (1. normal point; 2. outlier point)}

\item{geneID}{A geneID, which should be plotted. Can also be a vector.}

\item{global}{Flag to plot a global Q-Q plot, default FALSE}

\item{samplePoints}{Sample points for Q-Q plot, defaults to max 30k points}

\item{legendPos}{Set legendpos, by default topleft.}

\item{outlierRatio}{The fraction to be used for the outlier sample filtering}

\item{conf.alpha}{If set, a confidence interval is plotted, defaults to 0.05}

\item{xlim}{The x limits for the plot or NULL to use the full data range}

\item{ylim}{The y limits for the plot or NULL to use the full data range}

\item{normalized}{If TRUE, the normalized counts are used, the default,
otherwise the raw counts}

\item{log}{If TRUE, the default, counts are plotted in log10.}

\item{...}{Additional arguments passed to the 
\code{\link[gplots]{heatmap.2}} function}

\item{rowCentered}{If TRUE, the counts are row-wise (gene-wise) centered}

\item{rowCoFactor}{A vector of co-factors for color coding the rows}

\item{rowColSet}{A vector of colors or a color set from RColorBrewer}

\item{colCoFactor}{A vector of co-factors for color coding the columns}

\item{colColSet}{A vector of colors or a color set from RColorBrewer}

\item{nCluster}{An integer to be used for cutting the dendrogram into groups}

\item{annotateCluster}{If TRUE and nCluster is an integer the grouping 
is saved in the returned OutriderDataSet}

\item{dendrogram}{A character string indicating whether to draw 
'none', 'row', 'column' or 'both' dendrograms.}

\item{names}{character string indicating whether to draw 
'none', 'row', 'col', or 'both' names.}

\item{...}{Additional parameters passed to plot() or plot_ly() if not stated
otherwise in the details for each plot function}
}
\value{
An OutriderDataSet object

If base R graphics are used nothing is returned else the plotly or
            the gplot object is returned.
}
\description{
The OUTRIDER package provides mutliple functions to visualize 
the data and the results of a full data set analysis.

This is the list of all plotting function provided by OUTRIDER:
\itemize{
  \item plotVolcano()
  \item plotQQ()
  \item plotExpressionRank()
  \item plotCountCorHeatmap()
  \item plotAberrantPerSample()
  \item plotFPKM()
  \item plotDispEsts()
  \item plotPowerAnalysis()
}

For a detailed description of each plot function please see the details.
Most of the functions share the same parameters.
}
\details{
\code{plotVolcano}: the volcano plot is sample-centric. It plots for a given
sample the negative log10 nominal P-values against the Z-scores for all
genes.

\code{plotQQ}: the quantile-quantile plot for a given gene or if 
\code{global} is set to \code{TRUE} over the full data set. Here the 
observed P-values are plotted against the expected ones in the negative 
log10 space.

\code{plotExpressionRank}: This function plots for a given gene the 
expression level against the expression rank for all samples. This can 
be used with normalized and unnormalized expression values.

\code{plotCountCorHeatmap}: The correlation heatmap of the count data
of the full data set. Default the values are log transformed and 
row centered. This function returns an OutriderDataSet with annotated 
clusters if requested.
}
\examples{
ods <- makeExampleOutriderDataSet(dataset="Kremer")
ods <- filterExpression(ods, minCounts=TRUE)
ods <- OUTRIDER(ods)

plotVolcano(ods, 1)
plotVolcano(ods, 'MUC1404')

plotQQ(ods, 1)
plotQQ(ods, global=TRUE, outlierRatio=0.001)

plotExpressionRank(ods, 1)
plotExpressionRank(ods, "FAAH", normalized=FALSE, 
        log=FALSE, main="Over expression outlier")

sex <- sample(c("female", "male"), dim(ods)[2], replace=TRUE)
colData(ods)$sex <- sex
ods <- plotCountCorHeatmap(ods, annotateCluster=TRUE)
ods <- plotCountCorHeatmap(ods, colCoFactor="sex")
head(colData(ods)$clusterNumber)

}
