% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotPowerAnalysis}
\alias{plotPowerAnalysis}
\title{PowerAnalysis plot}
\usage{
plotPowerAnalysis(ods)
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}
}
\value{
The ggplot object
}
\description{
The power analysis plot should give the user a ruff estimate of the events
one can be detected with OUTRIDER. Based on the dispersion of the provided
OUTRIDER data set the theoretical P-value over the mean expression is
plotted. This is done for different expression levels. The curves are 
smooths to make the reading of the plot easier.
}
\examples{
ods <- makeExampleOutriderDataSet(100, 60)
ods <- estimateSizeFactors(ods)
ods <- fit(ods)

plotPowerAnalysis(ods)

}
