% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotQQ}
\alias{plotQQ}
\title{Plot a QQ-plot for a given gene}
\usage{
plotQQ(ods, geneID = 1:length(ods), global = FALSE, padj = 0.05,
  subset = 1:ncol(ods), main = NULL, maxOutlier = NULL, ...)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{geneID}{Gene to be plotted}

\item{global}{Flag to plot a global QQ-plot, default FALSE}

\item{padj}{significance level to mark outliers}

\item{subset}{samples to subset for}

\item{main}{title for the plot}

\item{maxOutlier}{y-axis range for the plot}
}
\value{
None
}
\description{
Plot a QQ-plot for a given gene
}
\examples{
ods <- makeExampleOutriderDataSet(20, 400)
ods <- OUTRIDER(ods)
plotQQ(ods, 1)

}
