% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectRversion.R
\name{predictC}
\alias{predictC}
\title{Predict controlled Counts.}
\usage{
predictC(w, k, s, xbar)
}
\arguments{
\item{w}{weight matrix}

\item{k}{counts}

\item{s}{size factors}

\item{xbar}{offset}
}
\value{
Returns the predicted corrections (predicted means).
}
\description{
Predict controlled Counts.
}
