% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUTRIDER.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing}
\usage{
preprocess(ods, prepro_options = getDefaultPreproParams(ods))
}
\arguments{
\item{ods}{Outrider2DataSet}

\item{prepro_options}{List of preprocessing options. The default settings 
can be obtained with getDefaultPreproParams(ods), and can be modified. 
Recognized options for this function are 'prepro_func' (preprocessing 
function, either NULL or a function), 'sf_norm' (TRUE or FALSE) and 
'data_trans' (transformation function, one of 'log', 'log1p', or NULL).}
}
\value{
An OutriderDataSet, with preprocessing output added.
}
\description{
Preprocesses an Outrider2DataSet according to the supplied parameter 
settings
}
\examples{
ods <- makeExampleOutriderDataSet()
prepro_opts <- getDefaultPreproParams(ods)
prepro_opts
ods <- preprocess(ods, prepro_options=prepro_opts)

}
