% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{writeNBModel}
\alias{writeNBModel}
\alias{readNBModel}
\title{Read/Write NB Model}
\usage{
writeNBModel(ods, modelFile)

readNBModel(ods, modelFile)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{modelFile}{specify file, where model should be saved.}
}
\description{
Writes out to a file or reads in from a file a fitted NB model.
The model contains:
  * feature name
  * log geometric means
  * fitted mean
  * dispersion
}
\examples{
# fit and write out model
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- fit(ods)
writeNBModel(ods, modelFile="model.tsv")

# read saved model
ods <- makeExampleOutriderDataSet()
ods <- readNBModel(ods, 'model.tsv')
ods <- estimateSizeFactors(ods)
ods <- computePvalues(ods)

}
