% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-results.R
\docType{methods}
\name{results}
\alias{results}
\alias{results}
\alias{results,OutriderDataSet-method}
\title{Accessor functions for the 'result' object in a OutriderDataSet
object.}
\usage{
results(object, ...)

\S4method{results}{OutriderDataSet}(object, padj = 0.05, zScore = 3,
  round = TRUE, all = FALSE)
}
\arguments{
\item{object}{OutriderDataSet}

\item{padj}{padj cutoff}

\item{zScore}{absolute Z-score cutoff}

\item{round}{rounding the results.}

\item{all}{By default FALSE, only significant read counts are listed in the 
results.}
}
\description{
This function extracts all results based on the given filter criteria.
}
\examples{
set.seed(42)
ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)
res <- results(ods)

res

}
