% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-gridSearch.R
\name{searchEncDim}
\alias{searchEncDim}
\title{SearchEncDim}
\usage{
searchEncDim(ods, params = seq(5, 30, 5), freq = 0.01, zScore = 3,
  inj = "both", BPPARAM = bpparam())
}
\arguments{
\item{ods}{OutriderDataSet}

\item{params}{set of possible q values.}

\item{freq}{frequncy of outlier by default 1E-2}

\item{zScore}{injection Z-score, by default 3.}

\item{inj}{injection strategy, by default 'both'.}

\item{BPPARAM}{BPPARAM object by default bpparam().}
}
\value{
optimal parameter value.
}
\description{
finds the optimal encoding dimension for a given data set.
}
\examples{
ods <- makeExampleDataSet
searchEncDim(ods)
}
