% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-evaluation.R
\name{aberrant}
\alias{aberrant}
\title{aberrant}
\usage{
aberrant(ods, padj = 0.05, zScore = 3, by = c("none", "sample", "gene"))
}
\arguments{
\item{ods}{a OutriderDataSet object}

\item{padj}{the padjusted cutoff}

\item{zScore}{the absolute Z-score cutoff, 
if NA or NULL no Z-score cutoff is used}

\item{by}{if the results should be summarized by 'sample', 
'gene' or not at all.}
}
\value{
the number of aberrent events by gene or sample or a TRUE/FALSE 
      matrix of the size sample x gene of aberrent events.
}
\description{
identifies the aberrant counts or returns the number of aberrant counts per
gene or sample.
}
\examples{
ods <- makeExampleOutriderDataSet(n=50)
ods <- OUTRIDER(ods)

aberrant(ods)[1:10,1:10]
aberrant(ods, by="sample")
aberrant(ods, by="gene")

}
