% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-evaluation.R
\name{aberrant}
\alias{aberrant}
\alias{aberrant,Outrider2DataSet-method}
\title{Number of aberrant events}
\usage{
aberrant(object, ...)

\S4method{aberrant}{Outrider2DataSet}(
  object,
  padjCutoff = 0.05,
  zScoreCutoff = 0,
  by = c("none", "sample", "feature")
)
}
\arguments{
\item{object}{An OutriderDataSet object}

\item{...}{Currently not in use.}

\item{padjCutoff}{The padjust cutoff}

\item{zScoreCutoff}{The absolute Z-score cutoff, 
if NA or NULL no Z-score cutoff is used}

\item{by}{if the results should be summarized by 'sample', 
'feature' or not at all (default).}
}
\value{
The number of aberrant events by feature (e.g. gene) or sample or 
            a TRUE/FALSE matrix of the size sample x gene of aberrant events.
}
\description{
Identifies the aberrant events and returns the number of aberrant counts per
feature (e.g. gene) or sample or returns a matrix indicating aberrant events.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods, implementation='pca')

aberrant(ods)[1:10,1:10]
tail(sort(aberrant(ods, by="sample")))
tail(sort(aberrant(ods, by="feature")))

}
