% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectRversion.R
\name{autoCorrect}
\alias{autoCorrect}
\title{Autoencoder function to correct for confounders.}
\usage{
autoCorrect(ods, q, theta = 25, implementation = c("R", "python", "PEER",
  "robustR", "cooksR", "robustRM1", "robustRTheta", "PEER_residual", "pca",
  "robustTheta", "debug", "mask25", "RobTheta200", "RobNoFTheta200"),
  BPPARAM = bpparam(), ...)
}
\arguments{
\item{ods}{An OutriderDataSet object}

\item{q}{The encoding dimensions}

\item{theta}{The dispersion parameter}

\item{implementation}{"R", the default, will use the R implementation or 
"python" to use the python/tensorflow experimental implementation}

\item{BPPARAM}{A \code{\link[BiocParallel]{BiocParallelParam}} instance
to be used for parallel computing.}

\item{...}{passed on to the autoencoder implementing method. In the case of 
the R implementation it is passed to the optim function.}
}
\value{
An ods object including the control factors
}
\description{
This is the wrapper function for the autoencoder implementation. 
It can be used to call the standard R implementation or the experimental
Python implementation.
}
\examples{
ods <- makeExampleOutriderDataSet()
\dontshow{
    ods <- ods[1:10,1:10]
}
ods <- estimateSizeFactors(ods)
ods <- autoCorrect(ods)

plotCountCorHeatmap(ods, normalized=FALSE)
plotCountCorHeatmap(ods, normalized=TRUE)

}
