% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectRversion.R
\name{autoCorrect}
\alias{autoCorrect}
\title{Autoencoder function to correct for co-founders.}
\usage{
autoCorrect(ods, q = 20, theta = 25, implementation = c("R", "python"),
  ...)
}
\arguments{
\item{ods}{an OutriderDataSet object}

\item{q}{the encoding dimensions}

\item{theta}{the dispersion parameter}

\item{implementation}{"R", the default will use the R implementation or 
"python" to use the python/tensorflow experimental implementation}

\item{...}{passed on to the autoencoder implementing method.}
}
\value{
An ods object including the control factors
}
\description{
This is the wrapper function for the autoencoder implementation. 
It can be used to call the standard R implementation or the experimental
Python implementation.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- autoCorrect(ods)

plotCountCorHeatmap(ods, normalized=FALSE)
plotCountCorHeatmap(ods, normalized=TRUE)

}
