% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectPython.R
\name{autoCorrectPython}
\alias{autoCorrectPython}
\title{autoCorrect}
\usage{
autoCorrectPython(ods, save = FALSE, predict = FALSE, epochs = 250,
  param_path = NULL, param_exp_name = NULL, verbose = FALSE,
  modelName = NULL, modelDirectory = NULL, seed = NULL)
}
\arguments{
\item{ods}{an OutriderDataSet object}

\item{save}{if TRUE, the fitted model will be saved to the given location.}

\item{predict}{run autoCorrect in predict mode, which means that the 
prediction is based on a given model. Need a prefitted model.}

\item{epochs}{The number of epochs used for the autoCorrection training.}

\item{param_path}{The directory where the optimal parameters are stored}

\item{param_exp_name}{Name of the parameter file}

\item{verbose}{if TRUE further information about the training/predicting
of the autoencoder is printed.}

\item{modelName}{Name of the model to read/write}

\item{modelDirectory}{The directory where the model is located 
or should be stored}

\item{seed}{if this is an integer it will be used to set the seed in the 
python runtime environment}
}
\value{
An ods object including the control factors
}
\description{
This is the wrapper function for the Python based autoCorrect method. 
It calls the autoencoder correction factor estimation.
}
