% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExpression.R
\name{computeGeneLength}
\alias{computeGeneLength}
\title{Computes for each gene based on the GTF file the exon length}
\usage{
computeGeneLength(ods, gtfFile, format = "gtf", mapping = NULL, ...)
}
\arguments{
\item{gtfFile}{Can be a gft file or an txDb object with annotation.}

\item{...}{further arguments to \code{makeTxDbFromGFF}}
}
\description{
Computes for each gene based on the GTF file the exon length
}
\examples{

ods <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
annotationFile <- system.file("extdata", "gencode.v19.genes.small.gtf.gz",
        package="OUTRIDER")
ods <- computeGeneLength(ods, annotationFile)

mcols(ods)['basepairs']
fpkm(ods)[1:10,1:10]

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(org.Hs.eg.db)
ods <- makeExampleOutriderDataSet(dataset="KremerNBaderSmall")
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
mapping <- select(org.Hs.eg.db, keys=keys(txdb, keytype = "GENEID"), 
        keytype="ENTREZID", columns=c("SYMBOL"))
ods <- computeGeneLength(ods, txdb, mapping=mapping)

mcols(ods)['basepairs']
fpkm(ods)[1:10,1:10]

}
