% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectRversion.R
\name{computeLatentSpace}
\alias{computeLatentSpace}
\title{Extracting the latent space}
\usage{
computeLatentSpace(ods)
}
\arguments{
\item{ods}{An OutriderDataSet}
}
\value{
A matrix containing the latent space determined by the autoencoder.
}
\description{
Extracts the latent space from the OutriderDataSet object 
determined by the autoencoder.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- autoCorrect(ods)
head(computeLatentSpace(ods))

}
