% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValMatrix.R
\docType{methods}
\name{computePvalues}
\alias{computePvalues}
\alias{computePvalues,OutriderDataSet-method}
\alias{computePvalues,Outrider2DataSet-method}
\title{Calculate P-values}
\usage{
computePvalues(object, ...)

\S4method{computePvalues}{Outrider2DataSet}(
  object,
  alternative = c("two.sided", "greater", "less"),
  method = "BY",
  BPPARAM = bpparam()
)
}
\arguments{
\item{object}{An OutriderDataSet}

\item{...}{additional params, currently not used.}

\item{alternative}{Can be one of "two.sided", "greater" or "less" to specify
the alternative hypothesis used to calculate the P-values,
defaults to "two.sided"}

\item{method}{Method used for multiple testing}

\item{BPPARAM}{Can be used to parallelize the computation, defaults to
bpparam()}
}
\value{
An OutriderDataSet object with computed nominal and adjusted P-values
}
\description{
This function computes the P-values based on the fitted negative binomial
model. It computes two matrices with the same dimension as the count matrix
(samples x genes), which contain the corresponding P-values and 
adjusted P-values of every count.
}
\examples{
ods <- makeExampleOutriderDataSet()
\dontshow{
    ods <- ods[1:10,1:10]
}
ods <- estimateSizeFactors(ods)
ods <- fit(ods)

ods <- computePvalues(ods)

assays(ods)[['pValue']][1:10,1:10]

}
\seealso{
p.adjust
}
