% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZscoreMatrix.R
\docType{methods}
\name{computeZscores}
\alias{computeZscores}
\alias{computeZscores,OutriderDataSet-method}
\title{Z score computation}
\usage{
computeZscores(ods, ...)

\S4method{computeZscores}{OutriderDataSet}(ods, ...)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{...}{Further arguments passed to \code{ZscoreMatrix}}
}
\value{
An OutriderDataSet containing the z score matrix ("zScore") and
    the log2 fold changes ("l2fc") as asasys.
}
\description{
Computes the z scores for every count in the matrix. 
The z score is defined in the log2 space as follows:
\ifelse{html}{
    \out{z<sub>ij</sub> = (l<sub>ij</sub> - mu<sub>j</sub><sup>l</sup>)/
            sigma<sub>j</sub><sup>l</sup>}}{
    \deqn{z_{ij} = \frac{l_{ij} - \mu_j^l}{\sigma_j^l}}},
where l is the log2 transformed normalized count and mu and sigma the 
mean and standard deviation for gene j, respectively.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- computeZscores(ods)

assays(ods)[['zScore']][1:10,1:10]
assays(ods)[["l2fc"]][1:10,1:10]

}
