% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlForConfounders.R
\name{estimateBestQ}
\alias{estimateBestQ}
\title{Estimation of Q}
\usage{
estimateBestQ(ods = NULL, zScores = NULL)
}
\arguments{
\item{ods}{An OutriderDataSet object}

\item{zScores}{A z-score matrix}
}
\value{
The estimated dimension of hidden confounders
}
\description{
Estimating the best q for the given data set by Optimal Hard Thresholding
}
\examples{
ods <- makeExampleOutriderDataSet()

estimateBestQ(ods)

}
