% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-gridSearch.R
\name{findEncodingDim}
\alias{findEncodingDim}
\title{Find the optimal encoding dimension}
\usage{
findEncodingDim(ods, params = seq(5, min(30, ncol(ods), nrow(ods)), 2),
  freq = 0.01, zScore = 3, inj = "both", BPPARAM = bpparam())
}
\arguments{
\item{ods}{An OutriderDataSet}

\item{params}{Set of possible q values.}

\item{freq}{Frequency of outlier, defaults to 1E-2}

\item{zScore}{Injection Z-score, defaults to 3.}

\item{inj}{Injection strategy, by default 'both'.}

\item{BPPARAM}{BPPARAM object by default bpparam().}
}
\value{
The optimal encoding dimension
}
\description{
Finds the optimal encoding dimension for a given data set by running a 
grid search based on the provided parameter set.
}
\examples{

ods <- makeExampleOutriderDataSet(180, 60)
ods <- findEncodingDim(ods, params=2:9)

# plot the results of the dimension search
metadata(ods)$encDimTable[plot(encodingDimension, evaluationLoss)]

}
