% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNB.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,OutriderDataSet-method}
\alias{fit,OutriderDataSet-method}
\title{Fit negative bionmial distribution}
\usage{
fit(object, ...)

\S4method{fit}{OutriderDataSet}(object, dropExtremeRank = FALSE,
  modelFile = NULL, BPPARAM = bpparam())
}
\arguments{
\item{object}{OutriderDataSet}

\item{...}{additional arguments, currently not used.}

\item{dropExtremeRank}{Default FALSE, should not be used anymore.}

\item{modelFile}{file name to save the estimated parameters}

\item{BPPARAM}{by default bpparam()}
}
\value{
An OutriderDataSet object with the fitted model. Accessible through:
            \code{mcols(ods)[,c('mu', 'disp')]}.
}
\description{
Fit a negative binomial (NB) distribution to the counts per gene
over all samples using the precomputed control factors.
}
\examples{
ods <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods <- estimateSizeFactors(ods)
ods <- fit(ods)

mcols(ods)[1:10,c('mu', 'disp')]

}
