% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNB.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,OutriderDataSet-method}
\title{Fit the negative binomial distribution}
\usage{
fit(object, ...)

\S4method{fit}{OutriderDataSet}(object, method = "OUTRIDER",
  BPPARAM = bpparam(), excludeMask)
}
\arguments{
\item{object}{An OutriderDataSet}

\item{...}{additional arguments, currently not used.}

\item{BPPARAM}{by default bpparam()}
}
\value{
An OutriderDataSet object with the fitted model. Accessible through:
            \code{mcols(ods)[,c('mu', 'theta')]}.
}
\description{
Fit a negative binomial (NB) distribution to the counts per gene
over all samples using if available the precomputed control factors.
If no normalization factors are provided only the sizeFactors are used.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- fit(ods)

mcols(ods)[1:10,c('mu', 'theta')]

}
