% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrectEDversion.R
\name{fitAutoencoder}
\alias{fitAutoencoder}
\title{Main autoencoder fit function}
\usage{
fitAutoencoder(ods, q, robust = TRUE, thetaRange = c(0.1, 250),
  convergence = 1e-05, loops = 15, pValCutoff = 0.01, minMu = 0.01,
  initialize = TRUE, noRobustLast = TRUE, control = list(),
  BPPARAM = bpparam(), ...)
}
\description{
Main autoencoder fit function
}
