% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExpression.R
\name{fpkm}
\alias{fpkm}
\alias{fpm}
\title{Calculate FPM and FPKM values}
\description{
This is the fpm and fpkm function from DESeq2. For more details see: 
\code{\link[DESeq2]{fpkm}} and \code{\link[DESeq2]{fpm}}
}
\examples{
ods <- makeExampleOutriderDataSet()
mcols(ods)['basepairs'] <- round(rnorm(nrow(ods), 1000, 500))

mcols(ods)['basepairs']
fpkm(ods)[1:10,1:10]
fpm(ods)[1:10,1:10]

}
\seealso{
\code{\link[DESeq2]{fpkm}} \code{\link[DESeq2]{fpm}}
}
