% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helperFunctions.R,
%   R/getNSetterFuns.R
\name{modelParams}
\alias{modelParams}
\alias{modelParams<-}
\alias{observed}
\alias{observed<-}
\alias{expected}
\alias{expected<-}
\alias{preprocessed}
\alias{preprocessed<-}
\alias{getBestQ}
\alias{getter_setter_functions}
\alias{zScore,}
\alias{pValue,}
\alias{padj,}
\alias{theta,}
\alias{dispersion,}
\alias{zScore}
\alias{pValue}
\alias{padj}
\alias{dispersions,OutriderDataSet-method}
\alias{dispersions}
\alias{theta}
\alias{modelParams,Outrider2DataSet-method}
\alias{modelParams<-,Outrider2DataSet-method}
\alias{observed,Outrider2DataSet-method}
\alias{observed<-,Outrider2DataSet-method}
\alias{observed<-,OutriderDataSet-method}
\alias{preprocessed,Outrider2DataSet-method}
\alias{preprocessed<-,Outrider2DataSet-method}
\alias{expected,Outrider2DataSet-method}
\alias{expected<-,Outrider2DataSet-method}
\title{Getter/Setter functions}
\usage{
modelParams(object, paramName)

modelParams(object, paramName) <- value

observed(object, normalized = FALSE, minE = 0, ...)

observed(object, ...) <- value

expected(object, ...)

expected(object, ...) <- value

preprocessed(object, ...)

preprocessed(object, ...) <- value

getBestQ(ods)

zScore(ods)

pValue(ods)

padj(ods)

\S4method{dispersions}{OutriderDataSet}(object, ...)

theta(ods)

\S4method{modelParams}{Outrider2DataSet}(object, paramName)

\S4method{modelParams}{Outrider2DataSet}(object, paramName) <- value

\S4method{observed}{Outrider2DataSet}(object, normalized = FALSE, minE = 0.5, ...)

\S4method{observed}{Outrider2DataSet}(object, ...) <- value

\S4method{observed}{OutriderDataSet}(object, ...) <- value

\S4method{preprocessed}{Outrider2DataSet}(object, normalized = FALSE)

\S4method{preprocessed}{Outrider2DataSet}(object, ...) <- value

\S4method{expected}{Outrider2DataSet}(object, ...)

\S4method{expected}{Outrider2DataSet}(object, ...) <- value
}
\arguments{
\item{...}{Further arguments passed on to the underlying assay function.}

\item{ods, object}{An OutriderDataSet object.}
}
\value{
A matrix or vector dependent on the type of data retrieved.
}
\description{
This is a collection of small accessor/setter functions for easy access to
the values within the OUTRIDER model.
}
\examples{
ods <- makeExampleOutriderDataSet(10, 10)
ods <- OUTRIDER(ods)

zScore(ods)
pValue(ods)
padj(ods)
theta(ods)
theta(ods) == 1/dispersions(ods)
getBestQ(ods)

}
\seealso{
\code{\link[DESeq2]{estimateDispersions}}
}
