% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OutriderDataSet.R
\name{makeExampleOutriderDataSet}
\alias{makeExampleOutriderDataSet}
\title{Create example data sets for OUTRIDER}
\usage{
makeExampleOutriderDataSet(n = 1000, m = 100, dataset = NULL, ...)
}
\arguments{
\item{n}{number of simulated genes}

\item{m}{number of simulated samples.}

\item{dataset}{here one can select from the two provided data sets. #TODO}

\item{...}{further arguments to \code{makeExampleDESeqDataSet}}
}
\description{
Creates an example data set from a file or a generates a random counts.
}
\examples{
# A generic dataset 
ods1 <- makeExampleOutriderDataSet()
ods1

# A generic dataset with specificed sample size
ods2 <- makeExampleOutriderDataSet(n=10, m=100)
ods2

# A subset of a real world dataset from GTEx 
ods3 <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods3

}
