% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OutriderDataSet.R
\name{makeExampleOutriderDataSet}
\alias{makeExampleOutriderDataSet}
\title{Create example data sets for OUTRIDER}
\usage{
makeExampleOutriderDataSet(n = 1000, m = 100, freq = 0.01, zScore = 6,
  inj = c("both", "low", "high"), dataset = c("none", "GTExSkinSmall",
  "KremerNBaderSmall"), ...)
}
\arguments{
\item{n}{Number of simulated genes}

\item{m}{Number of simulated samples}

\item{freq}{Frequency of in-silico outliers}

\item{zScore}{Absolute z score of in-silico outliers (default 6).}

\item{inj}{Determines whether counts are injected with the strategy 
('both', 'low', 'high'), default is 'both'.}

\item{dataset}{If "none", the default, an example data set is simulated. 
One can also use example data set included in the package by
specifying 'GTExSkinSmall' or 'KremerNBaderSmall'}

\item{...}{Further arguments to \code{makeExampleDESeqDataSet}}
}
\value{
An OutriderDataSet containing an example dataset. Depending on the
            parameters it is based on a real data set or it is simulated
}
\description{
Creates an example data set from a file or simulates a data set based 
on random counts following a negative binomial distribution with injected
outliers with a fixed z score away from the mean of the gene.
}
\examples{
# A generic dataset 
ods1 <- makeExampleOutriderDataSet()
ods1

# A generic dataset with specificed sample size and injection method
ods2 <- makeExampleOutriderDataSet(n=300, m=50, inj='low')
ods2

# A subset of a real world dataset from GTEx 
ods3 <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods3

}
