% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OutriderDataSet.R
\name{makeExampleOutriderDataSet}
\alias{makeExampleOutriderDataSet}
\title{Create example data sets for OUTRIDER}
\usage{
makeExampleOutriderDataSet(n = 1000, m = 100, freq = 0.01, zScore = 6,
  inj = "both", ..., dataset = c("none", "GTExSkinSmall",
  "KremerNBaderSmall"))
}
\arguments{
\item{n}{number of simulated genes}

\item{m}{number of simulated samples.}

\item{...}{further arguments to \code{makeExampleDESeqDataSet}}

\item{dataset}{here one can select from the two example data sets.
One of 'none', 'GTExSkinSmall', or 'KremerNBaderSmall'.}
}
\value{
An OutriderDataSet containing an example dataset. Depending on the
            parameters it is based on a real data set or on simulated counts.
}
\description{
Creates an example data set from a file or a generates a random counts.
}
\examples{
# A generic dataset 
ods1 <- makeExampleOutriderDataSet()
ods1

# A generic dataset with specificed sample size
ods2 <- makeExampleOutriderDataSet(n=10, m=100)
ods2

# A subset of a real world dataset from GTEx 
ods3 <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods3

}
