% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUTRIDER.R
\name{OUTRIDER}
\alias{OUTRIDER}
\title{The OUTRIDER function runs the default OUTRIDER pipline. 
Combinig the fit, the computation of zScores and pValues.
All computed values are returned as a OutriderDataSet object.}
\usage{
OUTRIDER(object, autoControl = TRUE, save = FALSE, acModelName = NULL,
  acModelDir = NULL, nbModelFile = NULL)
}
\arguments{
\item{object}{main OutriderDataSet object, which contains all the data}

\item{autoControl}{option, to run autoControl}
}
\value{
OutriderDataSet with all the computed values
}
\description{
The values can be accessed by: assays(ods)[['value']]
}
\examples{
ods <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods <- OUTRIDER(ods)
  
assays(ods)[['pValue']][1:10,1:10]
res <- results(ods)
plotVolcano(ods, 1)
  
}
