% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUTRIDER.R
\name{OUTRIDER}
\alias{OUTRIDER}
\title{OUTRIDER - Full analysis pipeline}
\usage{
OUTRIDER(object, autoCorrect = TRUE, nbModelFile = NULL)
}
\arguments{
\item{object}{main OutriderDataSet object, which contains all the data}

\item{autoCorrect}{if TRUE the raw read counts will be controled 
for confounders by the autoencoder}

\item{nbModelFile}{The file where the negative binomial model should be saved}
}
\value{
OutriderDataSet with all the computed values. The values can be 
            accessed by: \code{assays(ods)[['value']]}
}
\description{
The OUTRIDER function runs the default OUTRIDER pipline. 
Combinig the fit, the computation of zScores and pValues.
All computed values are returned as a OutriderDataSet object.
}
\examples{
ods <- makeExampleOutriderDataSet(dataset="GTExSkinSmall")
ods <- OUTRIDER(ods)

assays(ods)[['pValue']][1:10,1:10]
res <- results(ods)
plotVolcano(ods, 1)

}
