% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValMatrix.R
\name{padjMatrix}
\alias{padjMatrix}
\title{FDR correction
Function to correct the computed pValues using BY FDR correction.
This function is called by the computePvalues method.}
\usage{
padjMatrix(ods, method)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{method}{adjustment method, by default 'BH'}
}
\value{
OutriderDataSet containing adjusted pValues
}
\description{
FDR correction
Function to correct the computed pValues using BY FDR correction.
This function is called by the computePvalues method.
}
