% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotAberrantPerSample}
\alias{plotAberrantPerSample}
\alias{plotAberrantPerSamplePlotly}
\title{Plot aberrant events per sample}
\usage{
plotAberrantPerSample(ods, padj = 0.05, main = NULL, outlierRatio = 0.001,
  col = c("#fdb462", "grey"), yadjust = c(1.2, 1.2), labLine = c(3.5, 3),
  ylab = "#Aberrantly expressed genes", labCex = par()$cex, ...)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{padj}{adjusted pvalue cutoff.}

\item{outlierRatio}{The fraction to be used for the outlier sample filtering}

\item{...}{further arguments to \code{aberrant}}
}
\value{
None
}
\description{
Plot aberrant events per sample
}
\examples{

ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)

plotAberrantPerSample(ods)

}
