% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotAberrantPerSample}
\alias{plotAberrantPerSample}
\alias{plotAberrantPerSamplePlotly}
\title{Aberrant events per sample plot}
\usage{
plotAberrantPerSample(ods, padjCutoff = 0.05, zScoreCutoff = 0,
  main = NULL, outlierRatio = 0.001, col = brewer.pal(3, "Dark2")[c(1,
  2)], yadjust = c(1.2, 1.2), labLine = c(3.5, 3),
  ylab = "#Aberrantly expressed genes", labCex = par()$cex, ...)
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}

\item{padjCutoff}{Significance level to mark outliers}

\item{zScoreCutoff}{Z-score cutoff to mark outliers}

\item{main}{Title for the plot, can be NULL, which is the default}

\item{outlierRatio}{The fraction to be used for the outlier sample filtering}

\item{col}{Set color for the points. If set, it must be a character vector 
of length 2. (1. normal point; 2. outlier point)}

\item{yadjust}{Option to adjust position of Median and 90 percentile labels.}

\item{labLine}{Option to move axis labels}

\item{ylab}{The y axis label}

\item{labCex}{The label cex parameter}

\item{...}{Further arguments to \code{\link{aberrant}}}
}
\value{
None
}
\description{
Plot aberrant events per sample
}
\examples{

ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)

plotAberrantPerSample(ods)

}
