% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotAberrantPerSample}
\alias{plotAberrantPerSample}
\alias{plotAberrantPerSamplePlotly}
\title{Plot aberrant events per sample}
\usage{
plotAberrantPerSample(ods, padj = 0.05, ...)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{padj}{adjusted pvalue cutoff.}

\item{...}{further arguments to \code{aberrant}}
}
\description{
Plot aberrant events per sample
}
\examples{

ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)

plotAberrantPerSample(ods)

}
