% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotCountCorHeatmap}
\alias{plotCountCorHeatmap}
\title{Correlation heatmap}
\usage{
plotCountCorHeatmap(ods, normalized = TRUE, rowCentered = TRUE,
  rowCoFactor = NULL, rowColSet = "Set1", colCoFactor = NULL,
  colColSet = "Set2", nCluster = 4, main = "Count correlation heatmap",
  annotateCluster = TRUE, dendrogram = "both", basePlot = TRUE,
  names = c("both", "row", "col", "none"), ...)
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}

\item{normalized}{If TRUE, the normalized counts are used, the default}

\item{rowCentered}{If TRUE, the counts are row-wise (gene-wise) centered}

\item{rowCoFactor}{A vector of co-factors for color coding the rows}

\item{rowColSet}{A vector of colors or a color set from RColorBrewer}

\item{colCoFactor}{A vector of co-factors for color coding the columns}

\item{colColSet}{A vector of colors or a color set from RColorBrewer}

\item{nCluster}{An integer to be used for cutting the dendrogram into groups}

\item{main}{Title for the plot, can be NULL, which is the default}

\item{annotateCluster}{If TRUE and nCluster is an integer the grouping 
is saved in the returned OutriderDataSet}

\item{dendrogram}{A character string indicating whether to draw 
'none', 'row', 'column' or 'both' dendrograms.}

\item{basePlot}{if TRUE, use the R base plot version, else use the plotly 
framekwork, which is the default}

\item{names}{character string indicating whether to draw 
'none', 'row', 'col', or 'both' names.}

\item{...}{Additional arguments passed to the 
\code{\link[gplots]{heatmap.2}} function}
}
\value{
An OutriderDataSet object
}
\description{
Correlation heatmap of the count data of the given samples
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- plotCountCorHeatmap(ods, annotateCluster=TRUE)

sex <- sample(c("female", "male"), dim(ods)[2], replace=TRUE)
colData(ods)$sex <- sex 
plotCountCorHeatmap(ods, colCoFactor="sex")

}
