% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotCountCorHeatmap}
\alias{plotCountCorHeatmap}
\alias{plotCountCorHeatmapPlotly}
\title{Correlation heatmap}
\usage{
plotCountCorHeatmap(x, normalized = TRUE, rowCoFactor = NULL,
  rowColSet = "Set1", colCoFactor = NULL, colColSet = "Set2",
  nCluster = 4, main = "Count correlation heatmap",
  annotateCluster = TRUE, dendrogram = "both", basePlot = TRUE,
  names = c("both", "row", "col", "none"), ...)

plotCountCorHeatmapPlotly(x, normalized = TRUE, rowCoFactor = NULL,
  rowColSet = "Set1", colCoFactor = NULL, colColSet = "Set2",
  nCluster = 4, main = "Count correlation heatmap",
  annotateCluster = TRUE, dendrogram = "both", ...)
}
\arguments{
\item{x}{An OutriderDataSet object}

\item{normalized}{if TRUE, the normalized counts are used, default.}

\item{rowCoFactor}{a vector of co-factors for color coding the rows}

\item{rowColSet}{A vector of colors or a color set from RColorBrewer}

\item{colCoFactor}{A vector of co-factors for color coding the columns}

\item{colColSet}{A vector of colors or a color set from RColorBrewer}

\item{nCluster}{An integer to be used for cutting the dendrogram into groups}

\item{main}{The title for the plot}

\item{annotateCluster}{if TRUE and nCluster is a integer the grouping 
is saved in the returned OutriderDataSet}

\item{dendrogram}{character string indicating whether to draw 
'none', 'row', 'column' or 'both' dendrograms.}

\item{basePlot}{if TRUE the heatmap.2 function is used otherwise the
interactive version from ggplotly is used.}

\item{names}{character string indicating whether to draw 
'none', 'row', 'col', or 'both' names.}

\item{...}{additional arguments to the \code{heatmap.2} function}
}
\value{
OutriderDataSet object
}
\description{
Correlation heatmap of the count data of the given samples
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- estimateSizeFactors(ods)
ods <- plotCountCorHeatmap(ods, annotateCluster=TRUE)

sex <- sample(c("female", "male"), dim(ods)[2], replace=TRUE)
colData(ods)$sex <- sex 
plotCountCorHeatmap(ods, colCoFactor="sex")

}
