% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\docType{methods}
\name{plotDispEsts,OutriderDataSet-method}
\alias{plotDispEsts,OutriderDataSet-method}
\title{Dispersion estimation plot}
\usage{
\S4method{plotDispEsts}{OutriderDataSet}(object, compareDisp)
}
\arguments{
\item{object}{An OutriderDataSet object containing the fitted model}

\item{compareDisp}{If TRUE, the default, and if the autoCorrect normalization
was used it computes the dispersion without autoCorrect and 
plots it for comparison.}
}
\value{
None
}
\description{
Plotting the dispersion of the OutriderDataSet model against the normalized 
mean count.
}
\examples{
ods <- makeExampleOutriderDataSet(160, 60)
ods <- estimateSizeFactors(ods)
ods <- fit(ods)

plotDispEsts(ods)

}
