% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotExpressionRank}
\alias{plotExpressionRank}
\title{Expression rank plot}
\usage{
plotExpressionRank(ods, geneID, padjCutoff = 0.05, zScoreCutoff = 0,
  normalized = TRUE, basePlot = FALSE, main = NULL)
}
\arguments{
\item{ods}{A OUTRIDER data set.}

\item{geneID}{Id of gene wich should be plotted. 
If multiple gene ids are supplied mulitple plots will be made.}

\item{padjCutoff}{padj cutoff for coloring significant genes}

\item{zScoreCutoff}{Z-score cutoff for coloring significant genes}

\item{normalized}{if TRUE the normalized counts are used 
otherwise the raw counts}

\item{basePlot}{R base plot version.}

\item{main}{string passed to main (title) of the plot.}
}
\value{
None
}
\description{
Plot expression over expression rank per gene.
The plot can be used before and after fitting.
}
\examples{
ods <- makeExampleOutriderDataSet(10, 100)
ods <- OUTRIDER(ods)
plotExpressionRank(ods, 1)

}
