% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotExpressionRank}
\alias{plotExpressionRank}
\title{Expression rank plot}
\usage{
plotExpressionRank(ods, geneID, padjCutoff = 0.05, zScoreCutoff = 0,
  normalized = TRUE, basePlot = FALSE, main = NULL, log = TRUE,
  col = c("gray", "firebrick"), ...)
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}

\item{geneID}{A gene as character or index, which should be plotted.
It can also be a vector.}

\item{padjCutoff}{Significance level to mark outliers}

\item{zScoreCutoff}{Z-score cutoff to mark outliers}

\item{normalized}{If TRUE the normalized counts are used,
otherwise the raw counts}

\item{basePlot}{if TRUE, use the R base plot version, else use the plotly 
framekwork, which is the default}

\item{main}{Title for the plot, can be NULL, which is the default}

\item{log}{If TRUE, the default, counts are plotted in log10.}

\item{col}{Set color for the points. If set, it must be a character vector 
of length 2. (1. normal point; 2. outlier point)}

\item{...}{Additional parameters passed to plot() or plot_ly().}
}
\value{
None or a plotly object
}
\description{
Plot expression over expression rank per gene.
The plot can be used before and after fitting.
}
\examples{
ods <- makeExampleOutriderDataSet(1000, 100)
ods <- OUTRIDER(ods)
plotExpressionRank(ods, 1)
plotExpressionRank(ods, 1, normalized=FALSE, log=FALSE, main="1. Gene")

}
