% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotFPKM}
\alias{plotFPKM}
\title{Plot histogram of FPKM values}
\usage{
plotFPKM(ods)
}
\arguments{
\item{ods}{a OutriderDataSet object containing the reads and fpkm values}
}
\value{
a ggplot object containing the FPKM plot
}
\description{
Plot histogram of FPKM values
}
\examples{
ctsFile <- system.file('extdata', 'GTExSkinSmall.tsv', package='OUTRIDER')
ods <- OutriderDataSet(countData=read.table(ctsFile, check.names=FALSE))
annotation <- system.file('extdata', 'gencode.v19.genes.small.gtf.gz', 
        package='OUTRIDER')

# Filter, storing the fpkm values and not subsetting the ods object.
ods <- filterExpression(ods, annotation, filterGenes=FALSE, savefpkm=TRUE)

#Display the pre distribution of counts.
plotFPKM(ods)

}
