% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotPowerAnalysis}
\alias{plotPowerAnalysis}
\title{PowerAnalysis Plot}
\usage{
plotPowerAnalysis(ods)
}
\arguments{
\item{ods}{OutriderDataSet}
}
\value{
a plot showing, wich P-value can be theoretically reached for 
specified fractions of expression reads.
}
\description{
The spikes in the plot do come from the fraction not beeing integer valued,
but the negative binomial is only defined for integer values.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- fit(ods)
plotPowerAnalysis(ods)
}
