% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotQQ}
\alias{plotQQ}
\title{Plot a QQ-plot for a given gene}
\usage{
plotQQ(ods, geneID = NULL, global = FALSE, padjCutoff = 0.05,
  zScoreCutoff = 0, main = NULL, sample = TRUE, legendPos = "topleft",
  outlierRatio = 0.001, conf.alpha = 0.05, pch = 16, col = NULL, ...)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{geneID}{Gene to be plotted}

\item{global}{Flag to plot a global QQ-plot, default FALSE}

\item{padjCutoff}{significance level to mark outliers}

\item{zScoreCutoff}{Z-score cutoff to coloring outliers}

\item{main}{title for the plot}

\item{sample}{sample points for QQplot, only used, when global==TRUE.}

\item{legendPos}{set legendpos, by default topleft.}

\item{outlierRatio}{The fraction to be used for the outlier sample filtering}

\item{conf.alpha}{if set, a confidence interval is plotted}

\item{pch}{integer or character to be used for plotting points}

\item{col}{set color}

\item{...}{additional arguments for the internal plotQQ function.
This arguments are currently used for development.}
}
\value{
None
}
\description{
Plot a QQ-plot for a given gene
}
\examples{
ods <- makeExampleOutriderDataSet(20, 400)
ods <- OUTRIDER(ods)
plotQQ(ods, 1)
plotQQ(ods, global=TRUE, filterOutliers=TRUE)

}
