% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotQQ}
\alias{plotQQ}
\title{Q-Q plots}
\usage{
plotQQ(ods, geneID = NULL, global = FALSE, padjCutoff = 0.05,
  zScoreCutoff = 0, main = NULL, samplePoints = TRUE,
  legendPos = "topleft", outlierRatio = 0.001, conf.alpha = 0.05,
  pch = 16, xlim = NULL, ylim = NULL, col = NULL)
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}

\item{geneID}{A gene as character or index, which should be plotted.
It can also be a vector.}

\item{global}{Flag to plot a global Q-Q plot, default FALSE}

\item{padjCutoff}{Significance level to mark outliers}

\item{zScoreCutoff}{Z-score cutoff to mark outliers}

\item{main}{Title for the plot, can be NULL, which is the default}

\item{samplePoints}{Sample points for Q-Q plot, defaults to max 30k points}

\item{legendPos}{Set legendpos, by default topleft.}

\item{outlierRatio}{The fraction to be used for the outlier sample filtering}

\item{conf.alpha}{If set, a confidence interval is plotted}

\item{pch}{Integer or character to be used for plotting the points}

\item{xlim}{The x limits for the plot or NULL to use the full data range}

\item{ylim}{The y limits for the plot or NULL to use the full data range}

\item{col}{Set color for the points. If set, it must be a character vector 
of length 2. (1. normal point; 2. outlier point)}
}
\value{
None
}
\description{
Plot a Q-Q plot for a given gene or a Q-Q plot over the full data set.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)
plotQQ(ods, 1)
plotQQ(ods, global=TRUE, outlierRatio=0.001)

}
