% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotQQ}
\alias{plotQQ}
\title{Plot a QQ-plot for a given gene}
\usage{
plotQQ(ods, geneID, global = FALSE, padj = 0.05, zScore = 3,
  subset = 1:ncol(ods), main = NULL, maxOutlier = NULL, sample = TRUE,
  legendPos = "topleft", filterOutliers = FALSE, conf.alpha = 0.05,
  outlierRatio = 0.001, col = c("black", "#fdb462"), ...)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{geneID}{Gene to be plotted}

\item{global}{Flag to plot a global QQ-plot, default FALSE}

\item{padj}{significance level to mark outliers}

\item{zScore}{Z-score cutoff to coloring outliers}

\item{subset}{samples to subset for}

\item{main}{title for the plot}

\item{maxOutlier}{y-axis range for the plot}

\item{sample}{sample points for QQplot, only used, when global==TRUE.}

\item{legendPos}{set legendpos, by default topleft.}

\item{filterOutliers}{If TRUE, in the global QQ plot the full data set will
be compared against the for outlier sample filtered data set}

\item{conf.alpha}{if set, a confidence interval is plotted}

\item{outlierRatio}{The fraction to be used for the outlier sample filtering}

\item{col}{set color}

\item{...}{additional arguments for the internal plotQQ function.
This arguments are currently used for development.}
}
\value{
None
}
\description{
Plot a QQ-plot for a given gene
}
\examples{
ods <- makeExampleOutriderDataSet(20, 400)
ods <- OUTRIDER(ods)
plotQQ(ods, 1)
plotQQ(ods, global=TRUE, filterOutliers=TRUE)

}
