% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(ods, sampleID, padjCutoff = 0.05, zScoreCutoff = 0,
  main = NULL, basePlot = FALSE, col = c("gray", "firebrick"))
}
\arguments{
\item{ods}{OutriderDataSet}

\item{sampleID}{sample which should be plotted. 
Can also be a vector of samples.}

\item{padjCutoff}{padj cutoff}

\item{zScoreCutoff}{Z-score cutoff}

\item{main}{string passed to main (title) of the plot.}

\item{basePlot}{R base plot version of the plot.}

\item{col}{colors for the points in the form of c(non outliers, outliers)}
}
\value{
None
}
\description{
Volcano plot for a given sample over all genes.
}
\examples{
ods <- makeExampleOutriderDataSet(1000, 100)
ods <- OUTRIDER(ods)
plotVolcano(ods, 1)

}
