% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(ods, sampleID, padjCutoff = 0.05, zScoreCutoff = 0, pch = 16,
  main = NULL, basePlot = FALSE, col = c("gray", "firebrick"))
}
\arguments{
\item{ods}{An OutriderDataSet, which is used to extract the data for plotting}

\item{sampleID}{A sampleID, which should be plotted. Can also be a vector.}

\item{padjCutoff}{Significance level to mark outliers}

\item{zScoreCutoff}{Z-score cutoff to mark outliers}

\item{pch}{Integer or character to be used for plotting the points}

\item{main}{Title for the plot, can be NULL, which is the default}

\item{basePlot}{if TRUE, use the R base plot version, else use the plotly 
framekwork, which is the default}

\item{col}{Set color for the points. If set, it must be a character vector 
of length 2. (1. normal point; 2. outlier point)}
}
\value{
The plotly object or NULL if if base R is used
}
\description{
Volcano plot for a given sample over all genes.
}
\examples{
ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)
plotVolcano(ods, 1)

}
