% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(ods, sampleID, padjCut = 0.05, zScoreCut = 3,
  basePlot = FALSE)
}
\arguments{
\item{ods}{OutriderDataSet}

\item{sampleID}{sample which should be plotted. 
Can also be a vector of samples.}

\item{padjCut}{padj cutoff}

\item{zScoreCut}{Z-score cutoff}

\item{basePlot}{R base plot version of the plot.}
}
\value{
None
}
\description{
Volcano plot for a given sample over all genes.
}
\examples{
ods <- makeExampleOutriderDataSet(10, 100)
ods <- OUTRIDER(ods)
plotVolcano(ods, 1)

}
