% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-results.R
\docType{methods}
\name{results}
\alias{results}
\alias{results}
\alias{results,OutriderDataSet-method}
\title{Accessor functions for the 'result' object in a OutriderDataSet
object.}
\usage{
results(object, ...)

\S4method{results}{OutriderDataSet}(object, padjCutoff = 0.05,
  zScoreCutoff = 0, round = TRUE, all = FALSE)
}
\arguments{
\item{object}{OutriderDataSet}

\item{...}{additional arguments, currently not used}

\item{padjCutoff}{padj cutoff}

\item{zScoreCutoff}{absolute Z-score cutoff}

\item{round}{rounding the results.}

\item{all}{By default FALSE, only significant read counts are listed in the 
results.}
}
\value{
A data.table where each row is an outlier event and the columns
            contain additional information about this event. Eg padj, l2fc
}
\description{
This function extracts all results based on the given filter criteria.
}
\examples{
set.seed(42)
ods <- makeExampleOutriderDataSet()
ods <- OUTRIDER(ods)
res <- results(ods)

res

}
